chore: move Option tests into option module
